Update allows.txt
add paketobuildpacks/** image group into allows.txt.

This image is for the Spring Boot Maven plugin. That plugin allows the image to have multiple layers. When you build it for the first time and then change the codebase, it doesn’t need to rebuild all the layers—just the one you modified.

this paketobuildpacks/base it's plugin default container image.